V01/F01 updated
Sidebar, header, slider реализованы в отдельных компонентах
