+++ need-for-drive/src/App.vue
@@ -1,18 +1,5 @@
 <template>
-  <div id="app">
-    <vmain />
-  </div>
+  <router-view />
 </template>
-
-<script>
-import Vmain from "@/views/v-main.vue";
-
-export default {
-  name: "Home",
-  components: {
-    Vmain,
-  },
-};
-</script>
 
 <style lang="scss"></style>
