+++ need-for-drive/src/assets/styles.scss
@@ -1,4 +1,5 @@
 @import "./variables.scss";
+@import "./stylesorder.scss";
 
 @font-face { 
   font-family: 'Roboto'; 
@@ -198,11 +199,11 @@ svg {
 }
 
 .header__logo {
-  margin: 32px 0 0 64px;
+  margin: 32px 0 32px 64px;
 }
 
 .header__map {
-  margin: 32px 64px 0 0;
+  margin: 32px 64px 32px 0;
   display: flex;
   flex-direction: row;
   align-items: center;
@@ -517,7 +518,7 @@ svg {
   }
 
   .about__button {
-    width: 100%;
+    width: 100vw;
     border-radius: 0px;
     margin: 0;
   }
