+++ need-for-drive/src/assets/stylesorder.scss
@@ -1,0 +1,18 @@
+@import "./variables.scss";
+
+.order-wrp {
+    flex-basis: 100%;
+}
+
+.nav-list {
+    display: flex;
+    flex-direction: row;
+    justify-content: start;
+    list-style-type:none;
+    padding: 0%;
+    // margin: 10px 64px 10px 64px;
+}
+
+.nav-list > li {
+    margin: 10px 64px 10px 64px;
+}
