+++ need-for-drive/src/components/v-slider.vue
@@ -1,0 +1,70 @@
+<template>
+  <slider height="100%" animation="normal">
+    <slider-item
+      v-for="(slide, index) in sliderList"
+      :key="index"
+      :style="slide"
+    >
+      <div class="slider-about">
+        <p class="slider-title">
+          {{ slide.title }}
+        </p>
+        <p class="slider-text">
+          {{ slide.text }}
+        </p>
+        <button class="slider-button">Подробнее</button>
+      </div>
+    </slider-item>
+  </slider>
+</template>
+
+<script>
+import { Slider, SliderItem } from "vue-easy-slider";
+
+const BASE_SLIDER_ITEM = {
+  backgroundRepeat: "no-repeat",
+  backgroundSize: "cover",
+  width: "100%",
+  height: "100%",
+};
+
+export default {
+  name: "v-slider",
+  components: {
+    Slider,
+    SliderItem,
+  },
+  data() {
+    return {
+      sliderList: [
+        {
+          ...BASE_SLIDER_ITEM,
+          backgroundImage: `url(${require("../assets/slide_parking.jpg")})`,
+          title: "Бесплатная парковка",
+          text: "Оставляйте машину на платных городских парковках и разрешенных местах, не нарушая ПДД, а также в аэропортах.",
+        },
+        {
+          ...BASE_SLIDER_ITEM,
+          backgroundImage: `url(${require("../assets/slide_insurance.jpg")})`,
+          title: "Страховка",
+          text: "Полная страховка страховка автомобиля",
+        },
+        {
+          ...BASE_SLIDER_ITEM,
+          backgroundImage: `url(${require("../assets/slide_petrol.jpg")})`,
+          title: "Бензин",
+          text: "Полный бак на любой заправке города за наш счёт",
+        },
+        {
+          ...BASE_SLIDER_ITEM,
+          backgroundImage: `url(${require("../assets/slide_service.jpg")})`,
+          title: "Обслуживание",
+          text: "Автомобиль проходит еженедельное ТО",
+        },
+      ],
+    };
+  },
+};
+</script>
+
+<style lang="scss"></style>
