+++ need-for-drive/src/router/index.js
@@ -1,16 +1,28 @@
 // TO DO переключение
 import Vue from "vue";
 import VueRouter from "vue-router";
-import Home from "../views/Home.vue";
+// import Home from "../views/Home.vue";
+import Vmain from "../views/v-main.vue";
+import Vorder from "../views/v-order.vue";
 
 Vue.use(VueRouter);
 
 const routes = [
   {
     path: "/",
-    name: "Home",
-    component: Home,
+    name: "v-main",
+    component: Vmain,
   },
+  {
+    path: "/order",
+    name: "v-order",
+    component: Vorder,
+  },
+  // {
+  //   path: "/",
+  //   name: "Home",
+  //   component: Home,
+  // },
   {
     path: "/about",
     name: "About",
